move file_in_file_list to common utilities
